requirements.txt
streamlit
pandas
plotly
supabase
python-dotenv
streamlit-plotly-events>=2.0.0
plotly>=5.0.0
python-dotenv>=0.19.0
